Revert "Merge branch 'master' of https://github.com/conan-io/conan-center-index into package/libtomcrypt"
This reverts commit 2eb9a9dbd576abc8a26540415259042495932fa0, reversing
changes made to 4db2221368aa32ad4a49c23d9ff7e6c8b985d5df.

+++ recipes/sdl_image/all/conanfile.py
@@ -6,7 +6,7 @@ from conan.tools.files import copy, get, rmdir
 from conan.tools.scm import Version
 import os
 
-required_conan_version = ">=2"
+required_conan_version = ">=1.53.0"
 
 
 class SDLImageConan(ConanFile):
@@ -66,11 +66,19 @@ class SDLImageConan(ConanFile):
         "wic": False,
     }
 
+    def export_sources(self):
+        if Version(self.version) < "2.6":
+            copy(self, "CMakeLists.txt", src=self.recipe_folder, dst=self.export_sources_folder)
+
     def config_options(self):
         if self.settings.os == "Windows":
             del self.options.fPIC
         if not is_apple_os(self):
             del self.options.imageio
+        if Version(self.version) < "2.6":
+            del self.options.qoi
+            del self.options.with_avif
+            del self.options.with_jxl
         if self.settings.os != "Windows":
             del self.options.wic
 
@@ -79,22 +87,26 @@ class SDLImageConan(ConanFile):
             self.options.rm_safe("fPIC")
         self.settings.rm_safe("compiler.cppstd")
         self.settings.rm_safe("compiler.libcxx")
+        if self.options.shared:
+            # sdl static into sdl_image shared is not allowed
+            self.options["sdl"].shared = True
 
     def layout(self):
         cmake_layout(self, src_folder="src")
 
     def requirements(self):
-        self.requires("sdl/[^2.28]", transitive_headers=True)
+        # Headers are exposed https://github.com/conan-io/conan-center-index/pull/16167#issuecomment-1508347351
+        self.requires("sdl/2.28.3", transitive_headers=True)
         if self.options.with_libtiff:
             self.requires("libtiff/4.6.0")
         if self.options.with_libjpeg:
-            self.requires("libjpeg/[>=9e]")
+            self.requires("libjpeg/9e")
         if self.options.with_libpng:
-            self.requires("libpng/[>=1.6 <2]")
+            self.requires("libpng/1.6.40")
         if self.options.with_libwebp:
-            self.requires("libwebp/[>=1.3.2 <2]")
-        if self.options.with_avif:
-            self.requires("libavif/[>=1.0.1 <2]")
+            self.requires("libwebp/1.3.2")
+        if self.options.get_safe("with_avif"):
+            self.requires("libavif/1.0.1")
 
     def validate(self):
         if self.options.shared and not self.dependencies["sdl"].options.shared:
@@ -102,39 +114,57 @@ class SDLImageConan(ConanFile):
         # TODO: libjxl doesn't support conan v2(yet)
         if self.options.get_safe("with_jxl"):
             raise ConanInvalidConfiguration(f"{self.ref} doesn't support with_jxl (yet)")
-        if Version(self.version).major != Version(self.dependencies["sdl"].ref.version).major:
-            raise ConanInvalidConfiguration(f"{self.ref} and sdl must have the same major version")
 
     def build_requirements(self):
-        self.tool_requires("cmake/[>=3.16 <4]")
+        if Version(self.version) >= "2.6":
+            self.tool_requires("cmake/[>=3.16 <4]")
 
     def source(self):
         get(self, **self.conan_data["sources"][self.version], strip_root=True)
 
     def generate(self):
         tc = CMakeToolchain(self)
-        tc.cache_variables["SDL2IMAGE_VENDORED"] = False
-        tc.cache_variables["SDL2IMAGE_DEPS_SHARED"] = False
-        tc.cache_variables["SDL2IMAGE_SAMPLES"] = False
-        tc.cache_variables["SDL2IMAGE_AVIF"] = self.options.with_avif
-        tc.cache_variables["SDL2IMAGE_BMP"] = self.options.bmp
-        tc.cache_variables["SDL2IMAGE_GIF"] = self.options.gif
-        tc.cache_variables["SDL2IMAGE_JPG"] = self.options.with_libjpeg
-        tc.cache_variables["SDL2IMAGE_JXL"] = self.options.with_jxl
-        tc.cache_variables["SDL2IMAGE_LBM"] = self.options.lbm
-        tc.cache_variables["SDL2IMAGE_PCX"] = self.options.pcx
-        tc.cache_variables["SDL2IMAGE_PNG"] = self.options.with_libpng
-        tc.cache_variables["SDL2IMAGE_PNM"] = self.options.pnm
-        tc.cache_variables["SDL2IMAGE_QOI"] = self.options.qoi
-        tc.cache_variables["SDL2IMAGE_SVG"] = self.options.svg
-        tc.cache_variables["SDL2IMAGE_TGA"] = self.options.tga
-        tc.cache_variables["SDL2IMAGE_TIF"] = self.options.with_libtiff
-        tc.cache_variables["SDL2IMAGE_WEBP"] = self.options.with_libwebp
-        tc.cache_variables["SDL2IMAGE_XCF"] = self.options.xcf
-        tc.cache_variables["SDL2IMAGE_XPM"] = self.options.xpm
-        tc.cache_variables["SDL2IMAGE_XV"] = self.options.xv
-        tc.cache_variables["SDL2IMAGE_BACKEND_WIC"] = self.options.get_safe("wic")
-        tc.cache_variables["SDL2IMAGE_BACKEND_IMAGEIO"] = self.options.get_safe("imageio")
+        if Version(self.version) < "2.6":
+            tc.variables["SDL_IMAGE_SRC_DIR"] = self.source_folder.replace("\\", "/")
+            tc.variables["BMP"] = self.options.bmp
+            tc.variables["GIF"] = self.options.gif
+            tc.variables["IMAGEIO"] = self.options.get_safe("imageio")
+            tc.variables["JPG"] = self.options.with_libjpeg
+            tc.variables["LBM"] = self.options.lbm
+            tc.variables["PCX"] = self.options.pcx
+            tc.variables["PNG"] = self.options.with_libpng
+            tc.variables["PNM"] = self.options.pnm
+            tc.variables["SVG"] = self.options.svg
+            tc.variables["TGA"] = self.options.tga
+            tc.variables["TIF"] = self.options.with_libtiff
+            tc.variables["WEBP"] = self.options.with_libwebp
+            tc.variables["XCF"] = self.options.xcf
+            tc.variables["XPM"] = self.options.xpm
+            tc.variables["XV"] = self.options.xv
+            tc.variables["SDL_IS_SHARED"] = self.dependencies["sdl"].options.shared
+        else:
+            tc.variables["SDL2IMAGE_VENDORED"] = False
+            tc.variables["SDL2IMAGE_DEPS_SHARED"] = False
+            tc.variables["SDL2IMAGE_SAMPLES"] = False
+            tc.variables["SDL2IMAGE_AVIF"] = self.options.get_safe("with_avif")
+            tc.variables["SDL2IMAGE_BMP"] = self.options.bmp
+            tc.variables["SDL2IMAGE_GIF"] = self.options.gif
+            tc.variables["SDL2IMAGE_JPG"] = self.options.with_libjpeg
+            tc.variables["SDL2IMAGE_JXL"] = self.options.get_safe("with_jxl")
+            tc.variables["SDL2IMAGE_LBM"] = self.options.lbm
+            tc.variables["SDL2IMAGE_PCX"] = self.options.pcx
+            tc.variables["SDL2IMAGE_PNG"] = self.options.with_libpng
+            tc.variables["SDL2IMAGE_PNM"] = self.options.pnm
+            tc.variables["SDL2IMAGE_QOI"] = self.options.get_safe("qoi")
+            tc.variables["SDL2IMAGE_SVG"] = self.options.svg
+            tc.variables["SDL2IMAGE_TGA"] = self.options.tga
+            tc.variables["SDL2IMAGE_TIF"] = self.options.with_libtiff
+            tc.variables["SDL2IMAGE_WEBP"] = self.options.with_libwebp
+            tc.variables["SDL2IMAGE_XCF"] = self.options.xcf
+            tc.variables["SDL2IMAGE_XPM"] = self.options.xpm
+            tc.variables["SDL2IMAGE_XV"] = self.options.xv
+            tc.variables["SDL2IMAGE_BACKEND_WIC"] = self.options.get_safe("wic")
+            tc.variables["SDL2IMAGE_BACKEND_IMAGEIO"] = self.options.get_safe("imageio")
         tc.generate()
         cd = CMakeDeps(self)
         cd.generate()
@@ -142,15 +172,22 @@ class SDLImageConan(ConanFile):
     def build(self):
         rmdir(self, os.path.join(self.source_folder, "external"))
         cmake = CMake(self)
-        cmake.configure()
+        if Version(self.version) < "2.6":
+            cmake.configure(build_script_folder=os.path.join(self.source_folder, os.pardir))
+        else:
+            cmake.configure()
         cmake.build()
 
-        rmdir(self, os.path.join(self.package_folder, "lib", "cmake"))
-        rmdir(self, os.path.join(self.package_folder, "lib", "pkgconfig"))
-        rmdir(self, os.path.join(self.package_folder, "share"))
+        if Version(self.version) >= "2.6":
+            rmdir(self, os.path.join(self.package_folder, "lib", "cmake"))
+            rmdir(self, os.path.join(self.package_folder, "lib", "pkgconfig"))
+            rmdir(self, os.path.join(self.package_folder, "share"))
 
     def package(self):
-        copy(self, "LICENSE.txt", src=self.source_folder, dst=os.path.join(self.package_folder, "licenses"))
+        if Version(self.version) < "2.6":
+            copy(self, "COPYING.txt", src=self.source_folder, dst=os.path.join(self.package_folder, "licenses"))
+        else:
+            copy(self, "LICENSE.txt", src=self.source_folder, dst=os.path.join(self.package_folder, "licenses"))
         cmake = CMake(self)
         cmake.install()
 
@@ -162,10 +199,11 @@ class SDLImageConan(ConanFile):
 
     def package_info(self):
         lib_postfix = ""
-        if self.settings.compiler == "msvc" and not self.options.shared:
-            lib_postfix += "-static"
-        if self.settings.build_type == "Debug":
-            lib_postfix += "d"
+        if Version(self.version) >= "2.6":
+            if self.settings.os == "Windows" and not self.options.shared:
+                lib_postfix += "-static"
+            if self.settings.build_type == "Debug":
+                lib_postfix += "d"
 
         self.cpp_info.set_property("cmake_file_name", "SDL2_image")
         self.cpp_info.set_property("cmake_target_name", "SDL2_image::SDL2_image")
@@ -177,6 +215,12 @@ class SDLImageConan(ConanFile):
         self.cpp_info.components["_sdl_image"].includedirs.append(os.path.join("include", "SDL2"))
 
         # TODO: to remove in conan v2 once legacy generators removed
+        self.cpp_info.names["cmake_find_package"] = "SDL2_image"
+        self.cpp_info.names["cmake_find_package_multi"] = "SDL2_image"
+        self.cpp_info.names["pkg_config"] = "SDL2_image"
+        target_name = "SDL2_image" if self.options.shared else "SDL2_image-static"
+        self.cpp_info.components["_sdl_image"].names["cmake_find_package"] = target_name
+        self.cpp_info.components["_sdl_image"].names["cmake_find_package_multi"] = target_name
         self.cpp_info.components["_sdl_image"].set_property("cmake_target_name", "SDL2_image::SDL2_image")
         self.cpp_info.components["_sdl_image"].set_property("pkg_config_name", "SDL2_image")
         self.cpp_info.components["_sdl_image"].requires = ["sdl::sdl"]
@@ -188,7 +232,7 @@ class SDLImageConan(ConanFile):
             self.cpp_info.components["_sdl_image"].requires.append("libpng::libpng")
         if self.options.with_libwebp:
             self.cpp_info.components["_sdl_image"].requires.append("libwebp::libwebp")
-        if self.options.with_avif:
+        if self.options.get_safe("with_avif"):
             self.cpp_info.components["_sdl_image"].requires.append("libavif::libavif")
         if self.options.get_safe("imageio") and not self.options.shared:
             self.cpp_info.components["_sdl_image"].frameworks = [
